If the SWS command was used to stop logging, the string is written to a file with the same name as the .lid file, but with a .gps extension. Issue #145
