+++ mat/lid_data_file.py
@@ -1,23 +1,11 @@
 from mat.sensor_data_file import SensorDataFile
 from math import ceil
-from mat.utils import parse_tags, epoch
+from mat.utils import parse_tags, epoch, write_sws_file, consecutive_numbers
 from datetime import datetime
 import numpy as np
 
 
 STOP_WITH_STRING_MARKER = -258
-
-
-def consecutive_numbers(data, number, count):
-    c = 0
-    for i, val in enumerate(data):
-        if val == number:
-            c += 1
-        else:
-            c = 0
-        if c == count:
-            return i-count+1
-    return len(data)
 
 
 class LidDataFile(SensorDataFile):
@@ -54,6 +42,8 @@ class LidDataFile(SensorDataFile):
             count=(self.PAGE_SIZE-self.mini_header_length())//2)
         if i == self.n_pages()-1:
             stop_idx = consecutive_numbers(data, STOP_WITH_STRING_MARKER, 7)
+            write_sws_file(self._file_path.replace('.lid', '.gps'),
+                           data[stop_idx+7:])
             data = data[:stop_idx]
         return data
